requirements.txt
appdirs==1.4.3
click==6.7
Flask==0.12
Flask-SQLAlchemy==2.2
itsdangerous==0.24
Jinja2==2.9.5
MarkupSafe==1.0
mysql-connector-python==2.1.3
packaging==16.8
PyGithub==1.33
PyJWT==1.4.2
pyparsing==2.2.0
requests==2.13.0
six==1.10.0
slacker==0.9.42
SQLAlchemy==1.1.7
virtualenv==15.1.0
Werkzeug==0.12.1